dev-assistant/pyproject.toml
[project]
name = "dev-assistant"
version = "0.1.0"
description = "Developer Assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.3",
    "fastembed>=0.8.0",
    "litserve>=0.2.17",
    "llama-index>=0.14.12",
    "llama-index-embeddings-openai-like>=0.2.2",
    "llama-index-llms-openai-like>=0.5.3",
    "llama-index-postprocessor-jinaai-rerank>=0.5.0",
    "llama-index-retrievers-bm25>=0.6.5",
    "llama-index-vector-stores-qdrant>=0.9.1",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-exporter-otlp>=1.40.0",
    "opentelemetry-instrumentation-celery>=0.61b0",
    "opentelemetry-instrumentation-fastapi>=0.61b0",
    "opentelemetry-instrumentation-httpx>=0.61b0",
    "opentelemetry-sdk>=1.40.0",
    "qdrant-client>=1.16.2",
    "redis>=7.4.0",
    "tokenizers>=0.22.2",
    "tree-sitter-language-pack>=0.13.0",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "flake8>=7.3.0",
    "ipykernel>=7.2.0",
    "isort>=8.0.1",
    "pre-commit>=4.5.1",
]
